requirements.txt
time
pyserial
RPi.GPIO
asyncio
logging
json 
datetime
bleuio